Feedback Implementation
A full list of OCC's roster is return if no bus or class is selected.
Volunteers are now searchable by their trainings and languages spoken.
The search bar in the volunteer section can search using 4 identifiers:
first name,
last name,
languages,
trainings
In addition to this, there is a button solely to see trainings